BUG: need to use snaketool-utils for profile double-handling support
