+++ setup.py
@@ -37,7 +37,8 @@ setup(
         "snakemake>=7.14.0",
         "pyyaml>=6.0",
         "click>=8.1.3",
-        "metasnek>=0.0.5"
+        "metasnek>=0.0.5",
+        "snaketool-utils>=0.0.3",
     ],
     entry_points={"console_scripts": ["phables=phables.__main__:main"]},
     include_package_data=True,
